refactor: split canonical read-only classifiers
